[faas] Fixed clippy warnings (#176)
* fixed clippy warnings

* fixed review comments

+++ crates/dslab-faas/src/extra/arima_extra.rs
@@ -86,11 +86,11 @@ pub(crate) fn arima_forecast<F: Fn(usize, &mut T) -> f64, T: Rng>(
     // get orders
     let ar_order = match ar {
         Some(par) => par.len(),
-        None => 0 as usize,
+        None => 0,
     };
     let ma_order = match ma {
         Some(par) => par.len(),
-        None => 0 as usize,
+        None => 0,
     };
 
     // initialize forecast with noise
